+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 This project versioning adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Added
+- Connection and request timeouts could be specified also when creating RemoteWebDriver from existing session ID.
 
 ## 1.5.0 - 2017-11-15
 ### Changed